requirements.txt
absl-py==2.1.0
accelerate==0.32.1
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
ansible==10.2.0
ansible-core==2.17.2
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argcomplete==3.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
ascii-magic==2.3.0
astor==0.8.1
astroid==3.2.4
asttokens==2.4.1
async-timeout==4.0.3
asyncache==0.3.1
asyncssh==2.15.0
atpublic==5.0
attrs==23.2.0
autopep8==2.3.1
avalanche-lib==0.5.0
azure-core==1.30.2
azure-storage-blob==12.19.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.4.2
blinker==1.8.2
boto==2.49.0
boto3==1.34.56
botocore==1.34.144
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
clarabel==0.9.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
cmeel==0.53.3
colorama==0.4.6
configobj==5.0.8
confluent-kafka==2.5.0
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.8
cvxopt==1.3.2
cvxpy==1.5.2
cycler==0.12.1
daqp==0.5.1.post1
decorator==5.1.1
dictdiffer==0.9.0
diffusers==0.29.2
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docker==6.1.3
docker-pycreds==0.4.0
dpath==2.2.0
dulwich==0.22.1
dvc==3.52.0
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.4.0
ecos==2.0.14
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.110.0
fasteners==0.19
filelock==3.15.4
filetype==1.2.0
fire==0.6.0
flake8==7.1.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.53.1
forbiddenfruit==0.1.4
frozenlist==1.4.1
fsspec==2024.6.1
funcy==2.0
gcs-oauth2-boto-plugin==3.2
gdown==5.2.0
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.43
google-apitools==0.5.32
google-auth==2.17.0
google-auth-httplib2==0.2.0
google-reauth==0.1.1
GPUtil==1.4.0
grandalf==0.8
greenlet==3.0.3
grpcio==1.64.1
gsutil==5.30
gto==1.7.1
gTTS==2.5.1
gunicorn==22.0.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.11.0
hagrid==0.3.124
highspy==1.7.2
httpcore==1.0.5
httplib2==0.20.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.5
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.0.0
intel-openmp==2021.4.0
ipython==8.26.0
isodate==0.6.1
isort==5.13.2
itables==1.7.1
iterative-telemetry==0.0.8
itsdangerous==2.2.0
jax==0.4.20
jaxlib==0.4.20
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json-tricks==3.17.3
kafka-python==2.0.2
kiwisolver==1.4.5
kombu==5.3.7
kr8s==0.13.5
learn2learn==0.2.0
lightning-bolts==0.7.0
lightning-utilities==0.11.5
loguru==0.7.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.14
mkl==2021.4.0
ml-dtypes==0.4.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
names==0.3.0
networkx==3.2.1
nir==1.0.4
nirtorch==1.0
nltk==3.8.1
nni==3.0
numpy==1.24.4
nvidia-ml-py==12.555.43
oauth2client==4.1.3
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.6
osqp==0.6.7.post0
packaging==24.1
pandas==2.2.1
paramiko==3.4.0
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
piqp==0.4.1
platformdirs==3.11.0
prettytable==3.10.2
prompt_toolkit==3.0.47
protobuf==4.25.3
proxsuite==0.6.6
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybboxes==0.1.6
pycapnp==2.0.0b2
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.1
pydot==3.0.1
pyflakes==3.2.0
pygit2==1.15.1
Pygments==2.18.0
pygtrie==2.5.0
pylint==3.2.6
pymongo==4.6.2
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-box==7.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jsonpath==1.1.1
PythonWebHDFS==0.2.3
pytorch-lightning==1.9.5
pytorchcv==0.0.67
pytz==2024.1
pyu2f==0.1.5
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
qdldl==0.1.7.post4
qpalm==1.2.3
qpsolvers==4.3.2
qpth==0.0.17
quadprog==0.1.12
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
resolvelib==1.0.1
responses==0.25.3
RestrictedPython==7.0
result==0.16.1
retry_decorator==1.1.1
rich==13.7.1
roboflow==1.1.36
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
safetensors==0.4.3
sahi==0.11.18
schema==0.7.7
scikit-learn==1.5.1
scipy==1.14.0
scmrepo==3.3.6
scs==3.2.6
seaborn==0.13.2
semver==3.0.2
sentry-sdk==2.10.0
setproctitle==1.3.3
shapely==2.0.5
shellingham==1.5.4
sherlock==0.4.1
shortuuid==1.0.13
shtab==1.7.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snntorch==0.9.1
sounddevice==0.4.7
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
sqltrie==0.11.0
stack-data==0.6.3
starlette==0.36.3
syft==0.8.6
sympy==1.13.0
tabulate==0.9.0
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.13.0
torch==2.3.1
torch_geometric==2.5.3
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.42.4
typeguard==4.1.5
typer==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.2.59
ultralytics-thop==2.0.0
urllib3==2.2.2
uvicorn==0.27.1
vine==5.1.0
virtualenv==20.26.3
virtualenv-api==2.1.18
voluptuous==0.15.2
waitress==3.0.0
wandb==0.17.4
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
win32-setctime==1.1.0
WTForms==3.1.2
yarl==1.9.4
yolov5==7.0.13
zc.lockfile==3.0.post1
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
